requirements.txt
requests
pymongo[srv]
python-dotenv
flask
pytest
pydantic
langchain
openai